switching mesh index buffer to use UINT instead of UINT16

+++ FYP/FYP/Commons/Mesh.cpp
@@ -53,7 +53,7 @@ UploadBuffer<Vertex>* Mesh::GetVertexUploadBuffer()
 	return m_pVertexBuffer;
 }
 
-UploadBuffer<UINT16>* Mesh::GetIndexUploadBuffer()
+UploadBuffer<UINT>* Mesh::GetIndexUploadBuffer()
 {
 	return m_pIndexBuffer;
 }
